refactor: 기존 header 컴포넌트 분리
